requirements.txt
# Core AG2 Framework
# Note: autogen-agentchat is the modern package (formerly pyautogen)
autogen-agentchat
autogen-core
autogen-ext

# Streamlit UI
streamlit>=1.29.0
streamlit-aggrid==0.3.4.post3
streamlit-card==0.0.4

# Data Processing
pandas==2.1.4
numpy==1.26.2
openpyxl==3.1.2


# AI/ML Models
openai==1.6.1
anthropic==0.8.1
langchain==0.1.0
langchain-openai==0.0.2
chromadb==0.4.22
sentence-transformers==2.2.2

# Git Integration
GitPython==3.1.40
PyGithub==2.1.1

# Cloud Providers
boto3==1.34.19  # AWS
google-cloud-storage==2.14.0  # GCP
azure-storage-blob==12.19.0  # Azure

# Kubernetes & DevOps
kubernetes==28.1.0
docker==7.0.0
pyyaml==6.0.1

# Infrastructure as Code
python-terraform==0.10.1
jinja2==3.1.2

# Database
sqlalchemy==2.0.23
redis==5.0.1
pymongo==4.6.1

# API & Web
requests==2.31.0
httpx==0.25.2
fastapi==0.109.0
uvicorn==0.25.0

# Monitoring & Logging
prometheus-client==0.19.0
python-json-logger==2.0.7
colorlog==6.8.0

# Security
cryptography==41.0.7
python-dotenv==1.0.0
pydantic==2.5.3
pydantic-settings==2.1.0

# Testing
pytest==7.4.3
pytest-asyncio==0.23.2
pytest-cov==4.1.0
pytest-mock==3.12.0

# Code Quality
black==23.12.1
flake8==7.0.0
pylint==3.0.3
mypy==1.8.0
isort==5.13.2

# Utilities
python-dateutil==2.8.2
pytz==2023.3
click==8.1.7
rich==13.7.0
tqdm==4.66.1

# File Processing
pillow==10.1.0
python-magic==0.4.27

# Template Engine
chevron==0.14.0

# Task Queue
celery==5.3.4
redis==5.0.1

# Configuration Management
python-decouple==3.8
configparser==6.0.0

# Async Support
aiofiles==23.2.1
aiohttp==3.9.1

# Development Tools
ipython==8.19.0
jupyter==1.0.0
notebook==7.0.6
